NHS111/NHS111.Business.DOS/ServiceType/OnlineServiceTypeFilter.cs
using System.Collections.Generic;
using System.Linq;
using NHS111.Models.Models.Web.FromExternalServices;
using BusinessModels = NHS111.Models.Models.Business;

namespace NHS111.Business.DOS.Service
{
    public class OnlineServiceTypeFilter : IOnlineServiceTypeFilter
    {
        public List<BusinessModels.DosService> FilterUnknownTypes(List<BusinessModels.DosService> resultsToMap)
        {
            resultsToMap.RemoveAll(s => s.OnlineDOSServiceType == OnlineDOSServiceType.Unknown);
            return resultsToMap.ToList();
        }

        public List<BusinessModels.DosService> FilterClosedCallbackServices(List<BusinessModels.DosService> resultsToMap)
        {
            resultsToMap.RemoveAll(s => s.OnlineDOSServiceType == OnlineDOSServiceType.Callback && !s.IsOpen);
            return resultsToMap.ToList();
        }
    }

    public interface IOnlineServiceTypeFilter
    {
        List<BusinessModels.DosService> FilterUnknownTypes(List<BusinessModels.DosService> resultsToMap);
        List<BusinessModels.DosService> FilterClosedCallbackServices(List<BusinessModels.DosService> resultsToMap);
    }
}
